Add `assembly` classifier to uber-jar, publish normal jar
With this change two dependency jars will be produced instead
of the current one.

The current uber-jar will now have a `-assembly` suffix on the
aretefact. In place of the un-classified jar will be a "thin" jar;
one with just the `spark-excel` classes.

This change means end-users can avoid pulling in the uber-jar to
environments where it may not be suitable, especially considering
the uber-jar contains unshaded dependencies.

Resolves #852